renamed IMDREADER to IMD
- fix #36
- renamed module
- changed imports in tests
- changed docs and examples
